HttpHeaderFormatter tests
